// src/types.ts
export interface Auction {
  auctionBeginTime: number
  auctionEndTime: number
  auctionStartingPrice: string
  authorAvatar: string
  authorId: number
  authorJoinTime: string
  authorNickName: string
  authorRealName: string
  category: null
  collectionId: number
  contractAddress: string
  createTime: string
  curAuctionChainOrderId: string
  curPendingId: number
  curSellChainOrderId: null
  curSellPrice: null
  highestBidderAvatar: null
  highestBidderId: number
  highestBidderNickName: null
  id: number
  imageUrl: string
  imgIpfsHash: null
  jsonIpfsHash: null
  lastTradePrice: string
  maxQuotedPrice: string
  miniImageUrl: string
  ownerAvatar: string
  ownerId: number
  ownerNickName: string
  remark: string
  royalties: string
  royaltiesPool: string
  saleBeginTime: null
  saleEndTime: null
  status: number
  title: string
  token: string
  tokenId: string
  tokenIdHex: string
  txHash: string
}